add geo_sigma2_flag for tor model
